+++ Sources/Paywall/Paywall/Paywall Presentation/SwiftUI/PaywallTrackModifier.swift
@@ -10,7 +10,7 @@ import Combine
 
 @available(iOS 13.0, *)
 struct PaywallTrackModifier: ViewModifier {
-  @Binding var shouldPresent: Bool
+  @Binding var shouldTrack: Bool
   @State private var programmaticallySetShouldPresent = false
   @State private var isInternallyPresenting = false
   var event: String
@@ -22,8 +22,8 @@ struct PaywallTrackModifier: ViewModifier {
 
   func body(content: Content) -> some View {
     content
-      .onReceive(Just(shouldPresent)) { _ in
-        updatePresentation(shouldPresent)
+      .onReceive(Just(shouldTrack)) { _ in
+        updatePresentation(shouldTrack)
       }
   }
 
@@ -49,24 +49,24 @@ struct PaywallTrackModifier: ViewModifier {
         onPresent: onPresent,
         onDismiss: { result in
           self.programmaticallySetShouldPresent = true
-          self.shouldPresent = false
+          self.shouldTrack = false
           self.isInternallyPresenting = false
           onDismiss?(result)
         },
         onSkip: { reason in
           self.programmaticallySetShouldPresent = true
-          self.shouldPresent = false
+          self.shouldTrack = false
           self.isInternallyPresenting = false
           onSkip?(reason)
         }
       )
     } else {
-      // When states change in SwiftUI views, the shouldPresent state seems to get temporarily
+      // When states change in SwiftUI views, the shouldTrack state seems to get temporarily
       // reset to false as it rerenders the view. This incorrectly calls Paywall.dismiss().
       // Also, when views get set up for the first time, Paywall.dismiss() was being called.
       // This guards against that.
       guard isInternallyPresenting else {
-        // This prevents Paywall.dismiss() being called when programmatically setting shouldPresent
+        // This prevents Paywall.dismiss() being called when programmatically setting shouldTrack
         // to false.
         if programmaticallySetShouldPresent {
           programmaticallySetShouldPresent = false