+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The changelog for `Paywall`. Also see the [releases](https://github.com/superwal
 - Mentions SwiftLint as a requirement to run the sample apps.
 - Deprecates `Paywall.debugMode`. All logs are now controlled by setting the paywall option `.logLevel`. The default `logLevel` is now `.warn`.
 - Fixes broken webview based deeplinks and closes the paywall view before calling the delegate handler.
+- Deprecates `Paywall.present` for `Paywall.trigger`.
 
 ---