+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The changelog for `Paywall`. Also see the [releases](https://github.com/superwal
 - Paywalls can now be configured via the dashboard to always present, regardless of the subscription status of the user.
 - Adds a `presentationStyleOverride` parameter to `Paywall.trigger()` and `Paywall.present()`. By setting this, you can override the configured presentation style on case by case basis.
 - Rules can now be limited by occurrence and date. For example, you could set a rule to only match 10 times within the last 5 hours.
+- Adds `Paywall.userId` to grab the id of the current user
 
 ### Fixes
 
@@ -40,6 +41,7 @@ The changelog for `Paywall`. Also see the [releases](https://github.com/superwal
 - Deprecates `Paywall.debugMode`. All logs are now controlled by setting the paywall option `.logLevel`. The default `logLevel` is now `.warn`.
 - Fixes broken webview based deeplinks and closes the paywall view before calling the delegate handler.
 - Deprecates `Paywall.present` for `Paywall.trigger`.
+- Fixes issue where preloaded paywalls would be cleared upon calling `Paywall.identify()` if config was called without a `userId`.
 
 ---