+++ CHANGELOG.md
@@ -22,7 +22,11 @@ The changelog for `Paywall`. Also see the [releases](https://github.com/superwal
 - Paywalls can now be configured via the dashboard to always present, regardless of the subscription status of the user.
 - Adds a `presentationStyleOverride` parameter to `Paywall.trigger()` and `Paywall.present()`. By setting this, you can override the configured presentation style on case by case basis.
 - Rules can now be limited by occurrence and date. For example, you could set a rule to only match 10 times within the last 5 hours.
-- Adds `Paywall.userId` to grab the id of the current user
+- Adds `Paywall.userId` to grab the id of the current user.
+- Adds `$url`, `$path`, `$pathExtension`, `$lastPathComponent`, `$host`, `$query`, `$fragment` as standard parameters to the `deepLink_open` event trigger (automatically tracked).
+- Parses URL parameters and adds them as trigger parameters to the `deepLink_open` event trigger (automatically tracked).
+- Fixes window logic for opening the debugger and launching paywalls on `deepLink_open`.
+- Launching a paywall using the `deepLink_open` Trigger now dismisses a currently presenting paywall before presenting the new one.
 
 ### Fixes
 
@@ -42,6 +46,7 @@ The changelog for `Paywall`. Also see the [releases](https://github.com/superwal
 - Fixes broken webview based deeplinks and closes the paywall view before calling the delegate handler.
 - Deprecates `Paywall.present` for `Paywall.trigger`.
 - Fixes issue where preloaded paywalls would be cleared upon calling `Paywall.identify()` if config was called without a `userId`.
+- Fixes logic for grabbing the active view controller.
 
 ---